convert item and location ID's to longs.

+++ src/main/java/gg/archipelago/aprandomizer/capability/data/WorldData.java
@@ -14,7 +14,7 @@ public class WorldData {
 
     private boolean jailPlayers = true;
 
-    private Set<Integer> locations = new HashSet<>();
+    private Set<Long> locations = new HashSet<>();
 
     public static final int DRAGON_KILLED = 30;
     public static final int DRAGON_SPAWNED = 20;
@@ -45,15 +45,15 @@ public class WorldData {
         this.jailPlayers = jailPlayers;
     }
 
-    public void addLocation(int location) {
+    public void addLocation(Long location) {
         this.locations.add(location);
     }
 
-    public void setLocations(int[] locations) {
+    public void setLocations(Long[] locations) {
         this.locations.addAll(Lists.newArrayList(Arrays.stream(locations).iterator()));
     }
 
-    public Set<Integer> getLocations() {
+    public Set<Long> getLocations() {
         return locations;
     }
 }
